requirements.txt
Flask~=3.1.3
python-dotenv~=1.2.2
yt_dlp
flask-cors
ytmusicapi~=1.11.5
requests~=2.32.5
mutagen~=1.47.0